requirements.txt
flask
tensorflow-macos
tensorflow-metal
pillow
numpy 